Clarify separation between git and deployment
There are two workflows here; one is a git workflow for making changes.
The other is a deployment workflow for advancing code to different
deployment tiers.

By making the end target of the git workflow "master" and the end target
of the deployment workflow "production", a separation is better
established in people's minds when they read through this.